Move notification-sidebar to shared/notification folder
