+++ src/main/webapp/app/shared/notification/notification-sidebar/notification-sidebar.scss
